+++ src/router/glib20/libglib/gio/gsettingsbackend.c
@@ -82,8 +82,8 @@ static gboolean g_settings_has_backend;
 static gboolean
 is_key (const gchar *key)
 {
-  gint length;
-  gint i;
+  size_t length;
+  size_t i;
 
   g_return_val_if_fail (key != NULL, FALSE);
   g_return_val_if_fail (key[0] == '/', FALSE);
@@ -101,8 +101,8 @@ is_key (const gchar *key)
 static gboolean
 is_path (const gchar *path)
 {
-  gint length;
-  gint i;
+  size_t length;
+  size_t i;
 
   g_return_val_if_fail (path != NULL, FALSE);
   g_return_val_if_fail (path[0] == '/', FALSE);
@@ -403,7 +403,7 @@ g_settings_backend_changed (GSettingsBackend *backend,
  * end with '/' and must not contain '//').
  *
  * The meaning of this signal is that any of the key names resulting
- * from the contatenation of @path with each item in @items may have
+ * from the concatenation of @path with each item in @items may have
  * changed.
  *
  * The same rules for when notifications must occur apply as per
