+++ src/router/glib20/libglib/gobject/gbinding.c
@@ -256,8 +256,8 @@ struct _GBinding
 
   GBindingFlags flags;
 
-  guint source_notify; /* LOCK: unbind_lock */
-  guint target_notify; /* LOCK: unbind_lock */
+  unsigned long source_notify; /* LOCK: unbind_lock */
+  unsigned long target_notify; /* LOCK: unbind_lock */
   gboolean target_weak_notify_installed; /* LOCK: unbind_lock */
 
   /* a guard, to avoid loops */
@@ -1250,7 +1250,7 @@ g_object_bind_property_full (gpointer               source,
   if ((flags & G_BINDING_INVERT_BOOLEAN) &&
       (transform_to != NULL || transform_from != NULL))
     {
-      flags &= ~G_BINDING_INVERT_BOOLEAN;
+      flags &= (unsigned) ~G_BINDING_INVERT_BOOLEAN;
     }
 
   pspec = g_object_class_find_property (G_OBJECT_GET_CLASS (source), source_property);
